Standardize start_requests for singing spiders
* All spiders allow csv overrides
* Uses allowed_domains (which needed to be fixed for phillysacredharp)
