+++ phillysacredharp/phillysacredharp/spiders/singing.py
@@ -8,38 +8,21 @@ from spider_base import SpiderBase
 
 class SingingSpider(SpiderBase):
     name = "singing"
-    allowed_domains = ["phillysacredharp.com"]
+    allowed_domains = ["phillysacredharp.org"]
 
     def start_requests(self):
-        conn = self.open_db()
-        curs = conn.cursor()
-
-        curs.execute('SELECT audio_url FROM minutes WHERE audio_url LIKE \'%phillysacredharp.org%\'')
-        for row in curs.fetchall():
-            if not self.parse_file(row[0]):
-                for url in row[0].split(','):
-                    # Keystone recordings are all on the same page, so we split
-                    # them up by using a url hash. Because these will resolve to
-                    # the same url and scrapy does duplicate filtering, we need
-                    # to set dont_filter
-                    request = scrapy.Request(url, dont_filter=True)
-                    request.meta['original_url'] = url
-                    m = re.search(r'#(\d+)$', url)
-                    if m:
-                        request.meta['year'] = m.group(1)
-                    yield request
-
-        curs.close()
-        conn.close()
-
-    def parse_file(self, url):
-        m = re.search(r'phillysacredharp\.org/([^/]+)', url)
-        print "looking for file: ", m.group(1)
-        if m:
-            filename = m.group(1) + '.csv'
-            if os.path.isfile(filename):
-                self.parse_csv(filename, url)
-                return True
+        for url in self.get_audio_urls():
+            if not self.parse_file(url):
+                # Keystone recordings are all on the same page, so we split
+                # them up by using a url hash. Because these will resolve to
+                # the same url and scrapy does duplicate filtering, we need
+                # to set dont_filter
+                request = scrapy.Request(url, dont_filter=True)
+                request.meta['original_url'] = url
+                m = re.search(r'#(\d+)$', url)
+                if m:
+                    request.meta['year'] = m.group(1)
+                yield request
 
     def parse(self, response):
         year = response.meta.get('year')
